update 20.06.21 Morning

+++ admin/users.php
@@ -78,9 +78,11 @@ if (!isLoggedIn()) {
 						</td>
 						<td><?php echo $admin['role']; ?></td>
 						<td>
+							<?php if ($_SESSION['user']['username'] === 'admin'):?>
 							<a class="fa fa-pencil btn edit"
 								href="users.php?edit-admin=<?php echo $admin['id'] ?>">
 							</a>
+							<?php endif ?>
 						</td>
 					</tr>
 					<?php endif ?>
@@ -112,9 +114,11 @@ if (!isLoggedIn()) {
 								</a>
 							</td>
 							<td>
+							<?php if ($_SESSION['user']['username'] === 'admin'):?>
 								<a class="fa fa-trash btn delete" 
 									href="users.php?delete-admin=<?php echo $admin['id'] ?>">
 								</a>
+								<?php endif ?>
 							</td>
 						</tr>
 						<?php endif ?>
